% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ceRNAInteraction.R
\name{get_all_ceRNAInteractions}
\alias{get_all_ceRNAInteractions}
\title{Find All Possible ceRNA Interactions}
\usage{
get_all_ceRNAInteractions(
  disease_name = NULL,
  ensg_number = NULL,
  gene_symbol = NULL,
  gene_type = NULL,
  pValue = 0.05,
  pValueDirection = "<",
  mscor = NULL,
  mscorDirection = "<",
  correlation = NULL,
  correlationDirection = "<",
  sorting = NULL,
  descending = TRUE,
  limit = 100,
  offset = NULL,
  information = FALSE
)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search available.}

\item{ensg_number}{A vector of ensg number(s). If ensg number is set, gene symbol and gene type must be NULL. One of the three identifiers must be provided.}

\item{gene_symbol}{A vector of gene symbol(s). If gene symbol is set, ensg number and gene type must be NULL. One of the three identifiers must be provided.}

\item{gene_type}{Defines the type of gene of interest. One out of [3prime_overlapping_ncRNA, antisense, antisense_RNA, bidirectional_promoter_lncRNA, IG_C_gene, IG_C_pseudogene, IG_V_gene, IG_V_pseudogene, lincRNA, macro_lncRNA, miRNA, misc_RNA, Mt_rRNA, polymorphic_pseudogene, processed_pseudogene, processed_transcript, protein_coding, pseudogene, ribozyme, rRNA, rRNA_pseudogene, scaRNA, scRNA, sense_intronic, sense_overlapping, snoRNA, snRNA, TEC, TR_C_gene, TR_V_gene, TR_V_pseudogene, transcribed_processed_pseudogene, transcribed_unitary_pseudogene, transcribed_unprocessed_pseudogene, translated_processed_pseudogene, unitary_pseudogene, unprocessed_pseudogene, vaultRNA].}

\item{pValue}{Threshold of the FDR adjusted p-value. Default is 0.05.}

\item{pValueDirection}{Direction of the FDR adjusted p-value threshold (<, >). Must be set if pValue is set. Possible values are: "<", ">".}

\item{mscor}{Threshold of the 'multiple sensitivity correlation' (mscor).}

\item{mscorDirection}{Direction of the mscor threshold (<, >). Must be set if pValue is set. Possible values are: "<", ">".}

\item{correlation}{Threshold of the correlation.}

\item{correlationDirection}{Direction of the correlation threshold (<, >). Must be set if pValue is set. Possible values are: "<", ">".}

\item{sorting}{Possibilities for sorting of the results. Possible values are "pValue", "mscor" or "correlation".}

\item{descending}{Descending (TRUE, default) or ascending (FALSE) ordering of the results.}

\item{limit}{Number of results that should be shown. Default value is 100 and can be up to 1000.
For more results please use batches, the provided offset parameter or download the whole dataset.}

\item{offset}{Starting point from where results should be shown.}

\item{information}{All available information about genes displayed (TRUE) or just ensg number (FALSE, default).}
}
\value{
A data_frame containing all ceRNA interactions fitting the paramters.
}
\description{
Get all ceRNA interactions by given identifications (ensg_number, gene_symbol or gene_type), specific cancer type/dataset or different filter possibilities according different statistical values (e.g. FDR adjusted p-value).
}
\examples{
# Retrieve all possible ceRNAs for gene, identified by ensg_number,
# and threshold for pValue and mscor.
get_all_ceRNAInteractions(ensg_number=c("ENSG00000259090","ENSG00000217289"),
                          pValue=0.5, pValueDirection="<",
                          mscor=0.006, mscorDirection="<",
                          limit=15, information=FALSE)
}
