% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ceRNA.R
\name{get_ceRNA}
\alias{get_ceRNA}
\title{Find Specific ceRNA (Gene)}
\usage{
get_ceRNA(
  disease_name,
  gene_type = NULL,
  minBetweenness = NULL,
  minNodeDegree = NULL,
  minEigenvector = NULL,
  sorting = NULL,
  descending = TRUE,
  limit = 100,
  offset = NULL
)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search available.}

\item{gene_type}{Defines the type of gene of interest. One out of [3prime_overlapping_ncRNA, antisense, antisense_RNA, bidirectional_promoter_lncRNA, IG_C_gene, IG_C_pseudogene, IG_V_gene, IG_V_pseudogene, lincRNA, macro_lncRNA, miRNA, misc_RNA, Mt_rRNA, polymorphic_pseudogene, processed_pseudogene, processed_transcript, protein_coding, pseudogene, ribozyme, rRNA, rRNA_pseudogene, scaRNA, scRNA, sense_intronic, sense_overlapping, snoRNA, snRNA, TEC, TR_C_gene, TR_V_gene, TR_V_pseudogene, transcribed_processed_pseudogene, transcribed_unitary_pseudogene, transcribed_unprocessed_pseudogene, translated_processed_pseudogene, unitary_pseudogene, unprocessed_pseudogene, vaultRNA].}

\item{minBetweenness}{Threshold of the betweenness.}

\item{minNodeDegree}{Threshold of the degree.}

\item{minEigenvector}{Threshold of the eigenvektor.}

\item{sorting}{Possibilities for sorting of the results. Possible values are "pValue", "mscor" or "correlation".}

\item{descending}{Descending (TRUE, default) or ascending (FALSE) ordering of the results.}

\item{limit}{Number of results that should be shown. Default value is 100 and can be up to 1000.
For more results please use batches, the provided offset parameter or download the whole dataset.}

\item{offset}{Starting point from where results should be shown.}
}
\value{
A data_frame containing all ceRNAs (genes) satisfying the given filters.
}
\description{
Get all ceRNAs in a disease of interest (search not for a specific ceRNA, but search for all ceRNAs satisfying filter functions).
}
\examples{
get_ceRNA(disease_name = "kidney clear cell carcinoma",
          gene_type = "lincRNA", minBetweenness = 0.8)
}
