% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_datasetInformation}
\alias{get_datasetInformation}
\title{Dataset Information Function}
\usage{
get_datasetInformation(disease_name = NULL)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available (e.g. "kidney clear cell carcinoma" or just "kidney").}
}
\value{
Information about all or specific dataset as data_frame.
}
\description{
Get information about all available datasets to start browsing or search for a specific cancer type/dataset.
}
\examples{
get_datasetInformation("kidney clear cell carcinoma")

\dontrun{
get_datasetInformation(c("kidney clear cell carcinoma", "kidney papillary cell carcinoma"))
}
}
