% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GeneCount.R
\name{get_geneCount}
\alias{get_geneCount}
\title{Number of Times Gene Involved in Complete Network and Significant Interactions.}
\usage{
get_geneCount(
  disease_name = NULL,
  ensg_number = NULL,
  gene_symbol = NULL,
  minCountAll = NULL,
  minCountSign = NULL
)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available.}

\item{ensg_number}{A vector of ensg number(s). If ensg_number is set, gene_symbol must be NULL.}

\item{gene_symbol}{A vector of gene symbol(s). If gene_symbol is set, ensg_number must be NULL.}

\item{minCountAll}{Defines the minimal number of times a gene has to be involved in the complete network (e.g. the degree of the corresponding node must be greater than minCountAll).}

\item{minCountSign}{Defines the minimal number of times a gene has to be involved in significant (p.adj < 0.05) interactions in the network.}
}
\value{
A data_frame cotaining the amount of times a gene is involved in the complete network (equals to degree), column count_all, and in significant (FDR adjusted pValue < 0.05) interactions of the network, column count_sign.
}
\description{
Number of Times Gene Involved in Complete Network and Significant Interactions.
}
\examples{
# Get all genes from specific cancer with a minimum number of 150 at significant ceRNA interactions
get_geneCount(disease_name = "kidney clear cell carinoma", minCountSign = 150)
}
