% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionValues.R
\name{get_geneExprValues}
\alias{get_geneExprValues}
\title{Gene Expression Function}
\usage{
get_geneExprValues(disease_name, ensg_number = NULL, gene_symbol = NULL)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available (e.g. "kidney clear cell carcinoma" or just "kidney").}

\item{ensg_number}{A vector of ensg number(s). If ensg_number is set, gene_symbol must be NULL.}

\item{gene_symbol}{A vector of gene symbol(s). If gene_symbol is set, ensg_number must be NULL.}
}
\value{
A data_frame with gene expression values.
}
\description{
Get all expression values for gene(s) of interest.
}
\examples{
# Retrieve gene expression values for specif genes by ensg_numbers
get_geneExprValues(disease_name = "kidney clear cell carcinoma",
                   ensg_number = c("ENSG00000259090","ENSG00000217289"))

# Retrieve gene expression values for specif genes by gene_symbols
get_geneExprValues(disease_name = "kidney clear cell carcinoma",
                   gene_symbol = c("SEPT7P1","TIGAR"))

 \dontrun{
# Ensg_numbers and gene_symbols together.
get_geneExprValues(disease_name = "kidney clear cell carcinoma",
                   ensg_number = c("ENSG00000259090","ENSG00000217289"),
                   gene_symbol = c("SEPT7P1","TIGAR"))
}
}
