% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneOntology.R
\name{get_geneOntology}
\alias{get_geneOntology}
\title{Find Associated GO Terms}
\usage{
get_geneOntology(gene_symbol)
}
\arguments{
\item{gene_symbol}{A vector of gene symbol(s). Required parameter.}
}
\value{
A data_frame containing all associated GO terms.
}
\description{
Associated GO terms for gene(s) of interest. QuickGO - a fast web-based browser of the Gene Ontology and Gene Ontology annotation data - is used as external source.
}
\examples{
get_geneOntology(gene_symbol=c("PTEN","TIGAR"))
}
