% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hallmarks.R
\name{get_hallmark}
\alias{get_hallmark}
\title{Find Associated Cancer hallmarks.}
\usage{
get_hallmark(gene_symbol)
}
\arguments{
\item{gene_symbol}{A vector of gene symbol(s). Required parameter.}
}
\value{
A data_frame containing all associated cancer hallmarks.
}
\description{
Associated cancer hallmark for gene(s) of interest. Cancer Hallmark Genes (http://bio-bigdata.hrbmu.edu.cn/CHG/) is used as external source.
}
\examples{
get_hallmark(gene_symbol=c("PTEN"))
}
