% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurences.R
\name{get_miRNAOccurences}
\alias{get_miRNAOccurences}
\title{Occurences of Sponges miRNAs}
\usage{
get_miRNAOccurences(
  disease_name,
  mimat_number = NULL,
  hs_number = NULL,
  occurences = NULL,
  sorting = NULL,
  descending = TRUE,
  limit = 100,
  offset = NULL
)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search available.}

\item{mimat_number}{A vector of mimat_number(s). If mimat_number is set, hs_number must be NULL.}

\item{hs_number}{A vector of hs_number(s). If hs_number is set, mimat_number must be NULL.}

\item{occurences}{Threshold of amount of contributions a miRNA should be considered.}

\item{sorting}{Possibilities for sorting of the results. Possible values are "pValue", "mscor" or "correlation".}

\item{descending}{Descending (TRUE, default) or ascending (FALSE) ordering of the results.}

\item{limit}{Number of results that should be shown. Default value is 100 and can be up to 1000.
For more results please use batches, the provided offset parameter or download the whole dataset.}

\item{offset}{Starting point from where results should be shown.}
}
\value{
A data_frame with all miRNAs occuring at leat "occurences" times.
}
\description{
Get all mirna involved in cancer type/dataset of interest occuring a certain amout of times.
}
\examples{
get_miRNAOccurences(disease_name="kidney clear cell carcinoma", occurences = 1000, limit = 10)
}
