% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{get_overallCounts}
\alias{get_overallCounts}
\title{Overall Count for Interactions and miRNAs}
\usage{
get_overallCounts()
}
\value{
Overview of interaction counts about all or specific dataset as data_frame.
}
\description{
Function return current statistic about database - amount of shared miRNA, significant and insignificant interactions per dataset
}
\examples{
get_overallCounts()
}
