% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_miRNAInteraction.R
\name{get_specific_miRNAInteraction}
\alias{get_specific_miRNAInteraction}
\title{miRNA Induced ceRNA Interaction}
\usage{
get_specific_miRNAInteraction(
  disease_name = NULL,
  mimat_number = NULL,
  hs_number = NULL,
  pValue = 0.05,
  pValueDirection = "<",
  mscor = NULL,
  mscorDirection = "<",
  correlation = NULL,
  correlationDirection = "<",
  limit = 100,
  offset = NULL
)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search available.}

\item{mimat_number}{Mimat_number of interest. If mimat_number is set, hs_number must be NULL.}

\item{hs_number}{hs_number of interest. If hs_number is set, mimat_number must be NULL.}

\item{pValue}{Threshold of the FDR adjusted p-value. Default is 0.05.}

\item{pValueDirection}{Direction of the FDR adjusted p-value threshold (<, >). Must be set if pValue is set. Possible values are: "<", ">".}

\item{mscor}{Threshold of the 'multiple sensitivity correlation' (mscor).}

\item{mscorDirection}{Direction of the mscor threshold (<, >). Must be set if pValue is set. Possible values are: "<", ">".}

\item{correlation}{Threshold of the correlation.}

\item{correlationDirection}{Direction of the correlation threshold (<, >). Must be set if pValue is set. Possible values are: "<", ">".}

\item{limit}{Number of results that should be shown. Default value is 100 and can be up to 1000.
For more results please use batches, the provided offset parameter or download the whole dataset.}

\item{offset}{Starting point from where results should be shown.}
}
\value{
A data_frame containing all ceRNA interactions fitting the parameters.
}
\description{
Get all ceRNA interactions where miRNA(s) of interest (different identifiers available - e.g. hs number or mimat number) contribute to.
}
\examples{
# Retrieve all possible ceRNA interactions where miRNA of interest contribute to
get_specific_miRNAInteraction(disease_name = "kidney clear cell carcinoma",
                              mimat_number = "MIMAT0000076",
                              limit = 15)
\dontrun{
# Do not use both possible identifiers at once
get_specific_miRNAInteraction(disease_name = "kidney clear cell carcinoma",
                              mimat_number ="MIMAT0000076",
                              hs_number = "hsa-miR-21-5p",
                              limit = 15)
}
}
