% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{get_survAna_pValues}
\alias{get_survAna_pValues}
\title{Survival Analysis Log Rank pValues}
\usage{
get_survAna_pValues(disease_name, ensg_number = NULL, gene_symbol = NULL)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available.}

\item{ensg_number}{A vector of ensg number(s). If ensg_number is set, gene_symbol must be NULL.}

\item{gene_symbol}{A vector of gene symbol(s). If gene_symbol is set, ensg_number must be NULL.}
}
\value{
A data_frame with gene information and corresponding log rank test pValue. For raw data use function \code{\link[spongeWeb]{get_survAna_rates}}
}
\description{
Retrieve pValues from log rank test based on raw survival analysis data
}
\examples{
 # Retrieve gene expression values for specif genes by ensg_numbers
get_survAna_pValues(disease_name = "kidney clear cell carcinoma",
                    ensg_number = c("ENSG00000259090","ENSG00000217289"))

# Retrieve gene expression values for specif genes by gene_symbols
get_survAna_pValues(disease_name = "kidney clear cell carcinoma",
                    gene_symbol = c("SEPT7P1","TIGAR"))

 \dontrun{
# Ensg_numbers and gene_symbols together.
get_survAna_pValues(disease_name = "kidney clear cell carcinoma",
                    ensg_number = c("ENSG00000259090","ENSG00000217289"),
                    gene_symbol = c("SEPT7P1","TIGAR"))
}
}
