% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{get_survAna_rates}
\alias{get_survAna_rates}
\title{Survival Analysis Raw Data}
\usage{
get_survAna_rates(
  disease_name,
  ensg_number = NULL,
  gene_symbol = NULL,
  sample_ID = NULL
)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available.}

\item{ensg_number}{A vector of ensg number(s). If ensg_number is set, gene_symbol must be NULL. One of the two identifiers must be provided.}

\item{gene_symbol}{A vector of gene symbol(s). If gene_symbol is set, ensg_number must be NULL.}

\item{sample_ID}{A vector of sample_ID of the patient/sample of interest.}
}
\value{
A data_frame with gene and patient/sample information and the "group information" encoded by column "overexpressed".
Information about expression value of the gene (FALSE = underexpression, gene expression <= mean gene expression over all samples,
                                                TRUE = overexpression, gene expression >= mean gene expression over all samples)
}
\description{
Get all raw survival analysis data for kaplan meier plots.
}
\examples{
 # Retrieve survival rates for specif genes by ensg_numbers and specific patient/sample
get_survAna_rates(disease_name="kidney clear cell carcinoma",
                  ensg_number=c("ENSG00000259090", "ENSG00000217289"),
                  sample_ID = c("TCGA-BP-4968","TCGA-B8-A54F"))
 \dontrun{
# Ensg_numbers and gene_symbols together.
get_survAna_pValues(disease_name = "kidney clear cell carcinoma",
                    ensg_number = c("ENSG00000259090","ENSG00000217289"),
                    gene_symbol = c("SEPT7P1","TIGAR"))
}
}
