% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WikiPathwayKey.R
\name{get_WikiPathwayKey}
\alias{get_WikiPathwayKey}
\title{Find Associated wikipathway keys}
\usage{
get_WikiPathwayKey(gene_symbol)
}
\arguments{
\item{gene_symbol}{A vector of gene symbol(s). Required parameter.}
}
\value{
A data_frame containing all associated wikipathway keys.
}
\description{
Associated wikipathway keys for gene(s) of interest. The WikiPathways database is used as external source.
}
\examples{
get_WikiPathwayKey(gene_symbol=c("PTEN","GJA1"))
}
