% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionValues.R
\name{mirnaExprValues}
\alias{mirnaExprValues}
\title{miRNA Expression Function}
\usage{
mirnaExprValues(disease_name, mimat_number = NULL, hs_number = NULL)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available (e.g. "kidney clear cell carcinoma" or just "kidney").}

\item{mimat_number}{A vector of mimat_number(s). If mimat_number is set, hs_number must be NULL.}

\item{hs_number}{A vector of hs_number(s). If hs_number is set, mimat_number must be NULL.}
}
\value{
A data_table with mirna expression values.
}
\description{
Get all expression values for miRNA(s) of interest.
}
\examples{
# Retrieve gene expression values for specif miRNAs by mimat_numbers
mirnaExprValues(disease_name = "kidney clear cell carcinoma",
                mimat_number = c("MIMAT0000076", "MIMAT0000261"))

# Retrieve gene expression values for specif miRNAs by hs_numbers
mirnaExprValues(disease_name = "kidney clear cell carcinoma",
                hs_number = c("hsa-miR-21-5p", "hsa-miR-183-5p"))

 \dontrun{
# Mimat_numbers and hs_numbers together.
mirnaExprValues(disease_name = "kidney clear cell carcinoma",
                mimat_number = c("MIMAT0000076", "MIMAT0000261"),
                hs_number = c("hsa-miR-21-5p", "hsa-miR-183-5p"))
}
}
