% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{runInformation}
\alias{runInformation}
\title{Run Information Function}
\usage{
runInformation(disease_name)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset as string.
Fuzzy search is available (e.g. "kidney clear cell carcinoma" or just "kidney").}
}
\value{
Run information about dataset disease_name as data_frame.
}
\description{
Retrieve all used parameters of the SPONGE method to create published results for the cancer type/dataset of interest.
}
\examples{
runInformation("kidney clear cell carcinoma")

\dontrun{
runInformation(c("kidney clear cell carcinoma", "kidney papillary cell carcinoma"))
}
}
