% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{sillyPrinter}
\alias{sillyPrinter}
\title{Silly Printer function}
\usage{
sillyPrinter(r, x)
}
\arguments{
\item{r}{What you want in the second column.}

\item{x}{What you want in the first column.}
}
\value{
A tibble
}
\description{
Silly Printer function
}
\examples{
sillyPrinter(x = rnorm(5), r = rnorm(5))
}
