% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_miRNAInteraction.R
\name{specific_miRNAInteraction}
\alias{specific_miRNAInteraction}
\title{miRNA Induced ceRNA Interaction}
\usage{
specific_miRNAInteraction(
  disease_name = NULL,
  mimat_number = NULL,
  hs_number = NULL,
  limit = 100,
  offset = NULL,
  information = FALSE
)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search available.}

\item{mimat_number}{A vector of mimat_number(s). If mimat_number is set, hs_number must be NULL.}

\item{hs_number}{A vector of hs_number(s). If hs_number is set, mimat_number must be NULL.}

\item{limit}{Number of results that should be shown. Default value is 100 and can be up to 1000.
For more results please use batches, the provided offset parameter or download the whole dataset.}

\item{offset}{Starting point from where results should be shown.}

\item{information}{All available information about miRNA displayed (TRUE) or just ensg number (FALSE, default).}
}
\value{
A data_frame containing all ceRNA interactions fitting the paramters.
}
\description{
Get all ceRNA interactions where miRNA(s) of interest (different identifiers available - e.g. hs number or mimat number) contribute to.
}
\examples{
# Retrieve all possible ceRNA interactions where miRNA of interest contribute to
specific_miRNAInteraction(disease_name = "kidney clear cell carcinoma",
                          mimat_number = c("MIMAT0000076", "MIMAT0000261"),
                          limit = 15, information = FALSE)
\dontrun{
# Do not use both possible identifiers at once
  specific_miRNAInteraction(disease_name = "kidney clear cell carcinoma",
                          mimat_number =c("MIMAT0000076", "MIMAT0000261"),
                          hs_number = c("hsa-miR-21-5p", "hsa-miR-183-5p"),
                          limit = 15, information = FALSE)
}
}
