% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalAnalysis.R
\name{survAna_sampleInformation}
\alias{survAna_sampleInformation}
\title{Clinical Data For Samples/Patients}
\usage{
survAna_sampleInformation(disease_name, sample_ID = NULL)
}
\arguments{
\item{disease_name}{The name of the dataset of interest as string. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search is available.}

\item{sample_ID}{A vector of sample_ID of the patient/sample of interest.}
}
\value{
A data_frame with clinal date of all available or specific sample/patient.
}
\description{
Clinical Data For Samples/Patients
}
\examples{
survAna_sampleInformation(disease_name = "kidney clear cell carcinoma",
                          sample_ID = c("TCGA-BP-4968","TCGA-B8-A54F"))
\dontrun{
# Do not run function without disease.
survAna_sampleInformation(sample_ID = c("TCGA-BP-4968","TCGA-B8-A54F"))
}
}
