% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ceRNAInteraction.R
\name{get_specific_ceRNAInteractions}
\alias{get_specific_ceRNAInteractions}
\title{Find All Possible ceRNA Interactions Between Identifiers}
\usage{
get_specific_ceRNAInteractions(
  disease_name = NULL,
  ensg_number = NULL,
  gene_symbol = NULL,
  limit = 100,
  offset = NULL
)
}
\arguments{
\item{disease_name}{Name of the specific cancer type/dataset. If default is set, all available datasets with corresponding informations are shown.
Fuzzy search available.}

\item{ensg_number}{A vector of ensg number(s). If ensg number is set, gene symbol and gene type must be NULL. One of the three identifiers must be provided.}

\item{gene_symbol}{A vector of gene symbol(s). If gene symbol is set, ensg number and gene type must be NULL. One of the three identifiers must be provided.}

\item{limit}{Number of results that should be shown. Default value is 100 and can be up to 1000.
For more results please use batches, the provided offset parameter or download the whole dataset.}

\item{offset}{Starting point from where results should be shown.}
}
\value{
A data_frame containing all interactions between genes of interest.
}
\description{
Get all interactions between the given identifiers (ensg_number or gene_symbol).
}
\examples{
get_specific_ceRNAInteractions(disease_name = "kidney clear cell carcinoma",
                               ensg_number = c("ENSG00000259090","ENSG00000217289","ENSG00000152284"))
\dontrun{
# Do not use both identifiers at the same time
get_specific_ceRNAInteractions(disease_name = "kidney clear cell carcinoma",
                               ensg_number = c("ENSG00000259090","ENSG00000217289","ENSG00000152284"),
                               gene_symbol = c("SEPT7P1","TIGAR"))
}
}
